+++ opem/Test/test_Functions.py
@@ -44,6 +44,12 @@ False
 >>> filter_alpha({"alpha":-0.1})
 [Warning] Opem Automatically Set Alpha To Maximum Value (0)
 {'alpha': 0}
+>>> filter_range(0,100,0.1)
+[0, 100, 0.1]
+>>> filter_range(500,0,0.1)
+[0, 500, 0.1]
+>>> filter_range(500,0,-0.1)
+[0, 500, 0.1]
 >>> Input_dict=Get_Input({"T": "Cell Operation Temperature [K]", "PH2": "Partial Pressure [atm]", "PO2": "Partial Pressure [atm]"},input_item=input_test)
 >>> Input_keys=list(Input_dict.keys())
 >>> Input_keys.sort()
@@ -57,10 +63,10 @@ Test
 <BLANKLINE>
 <BLANKLINE>
 ###########
->>> check_update(1.1)
+>>> check_update(1.2)
 >>> check_update(0.1)
 ###########
-New Version (1.1) Is Available!
+New Version (1.2) Is Available!
 Website : http://www.ecsim.ir/opem
 ###########